+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/woopos/home/variations/WooPosVariationsViewModelTest.kt
@@ -16,8 +16,8 @@ import com.woocommerce.android.ui.woopos.home.WooPosChildrenToParentEventSender
 import com.woocommerce.android.ui.woopos.home.items.WooPosItemsViewModel
 import com.woocommerce.android.ui.woopos.home.items.WooPosPaginationState
 import com.woocommerce.android.ui.woopos.home.items.WooPosVariationsViewState
-import com.woocommerce.android.ui.woopos.home.items.variations.FetchResult
-import com.woocommerce.android.ui.woopos.home.items.variations.WooPosVariationsDataSource
+import com.woocommerce.android.ui.woopos.home.items.products.WooPosProductsDataSource
+import com.woocommerce.android.ui.woopos.home.items.products.WooPosProductsDataSource.VariationsResult
 import com.woocommerce.android.ui.woopos.home.items.variations.WooPosVariationsUIEvents
 import com.woocommerce.android.ui.woopos.home.items.variations.WooPosVariationsViewModel
 import com.woocommerce.android.ui.woopos.localcatalog.WooPosLocalCatalogSyncRepository
@@ -55,7 +55,7 @@ class WooPosVariationsViewModelTest {
     val coroutinesTestRule = WooPosCoroutineTestRule()
 
     private val getProductById: WooPosGetProductById = mock()
-    private val variationsDataSource: WooPosVariationsDataSource = mock()
+    private val variationsDataSource: WooPosProductsDataSource = mock()
     private val wooPosLocalCatalogM1Enabled: WooPosLocalCatalogM1Enabled = mock {
         on { invoke() } doReturn false // Default to using variationsDataSource
     }
@@ -124,8 +124,8 @@ class WooPosVariationsViewModelTest {
             ProductTestUtils.generateProductVariation(1, 1, "10.0").toWooPosVariation(mapper),
             ProductTestUtils.generateProductVariation(2, 1, "20.0").toWooPosVariation(mapper)
         )
-        whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
-            flowOf(FetchResult.Remote(Result.success(variations)))
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
+            flowOf(VariationsResult.Remote(Result.success(variations)))
         )
 
         // WHEN
@@ -147,8 +147,8 @@ class WooPosVariationsViewModelTest {
         val variations = listOf(
             ProductTestUtils.generateProductVariation(1, 1, "0.00").toWooPosVariation(mapper),
         )
-        whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
-            flowOf(FetchResult.Remote(Result.success(variations)))
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
+            flowOf(VariationsResult.Remote(Result.success(variations)))
         )
 
         // WHEN
@@ -170,8 +170,8 @@ class WooPosVariationsViewModelTest {
         val variations = listOf(
             ProductTestUtils.generateProductVariation(1, 1, "").toWooPosVariation(mapper),
         )
-        whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
-            flowOf(FetchResult.Remote(Result.success(variations)))
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
+            flowOf(VariationsResult.Remote(Result.success(variations)))
         )
 
         // WHEN
@@ -190,8 +190,8 @@ class WooPosVariationsViewModelTest {
     @Test
     fun `given empty variations list returned, when view model created, then view state is empty`() = runTest {
         // GIVEN
-        whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
-            flowOf(FetchResult.Remote(Result.success(emptyList())))
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
+            flowOf(VariationsResult.Remote(Result.success(emptyList())))
         )
 
         // WHEN
@@ -205,8 +205,8 @@ class WooPosVariationsViewModelTest {
     @Test
     fun `given error fetching variations, when view model created, then view state is error`() = runTest {
         // GIVEN
-        whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
-            flowOf(FetchResult.Remote(Result.failure(Exception("Fetch error"))))
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
+            flowOf(VariationsResult.Remote(Result.failure(Exception("Fetch error"))))
         )
 
         // WHEN
@@ -219,17 +219,17 @@ class WooPosVariationsViewModelTest {
     }
 
     @Test
-    fun `given variations, when pull to refresh triggered, then fetchFirstPage is called`() = runTest {
+    fun `given variations, when pull to refresh triggered, then fetchVariationsFirstPage() is called`() = runTest {
         // GIVEN
         val variations = listOf(
             ProductTestUtils.generateProductVariation(1, 1, "10.0").toWooPosVariation(mapper),
             ProductTestUtils.generateProductVariation(2, 1, "20.0").toWooPosVariation(mapper)
         )
-        whenever(variationsDataSource.fetchFirstPage(any(), eq(false))).thenReturn(
-            flowOf(FetchResult.Remote(Result.success(emptyList())))
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), eq(false))).thenReturn(
+            flowOf(VariationsResult.Remote(Result.success(emptyList())))
         )
-        whenever(variationsDataSource.fetchFirstPage(any(), eq(true))).thenReturn(
-            flowOf(FetchResult.Remote(Result.success(variations)))
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), eq(true))).thenReturn(
+            flowOf(VariationsResult.Remote(Result.success(variations)))
         )
 
         // WHEN
@@ -237,13 +237,13 @@ class WooPosVariationsViewModelTest {
         viewModel.onUIEvent(WooPosVariationsUIEvents.PullToRefreshTriggered(123L))
 
         // THEN
-        verify(variationsDataSource).fetchFirstPage(eq(123L), eq(true))
+        verify(variationsDataSource).fetchVariationsFirstPage(eq(123L), eq(true))
     }
 
     @Test
     fun `when pull to refresh triggered, then should track event`() = runTest {
-        whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
-            flowOf(FetchResult.Remote(Result.success(emptyList())))
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
+            flowOf(VariationsResult.Remote(Result.success(emptyList())))
         )
 
         // WHEN
@@ -268,11 +268,11 @@ class WooPosVariationsViewModelTest {
             ProductTestUtils.generateProductVariation(1, 1, "10.0").toWooPosVariation(mapper),
             ProductTestUtils.generateProductVariation(2, 1, "20.0").toWooPosVariation(mapper)
         )
-        whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
-            flowOf(FetchResult.Remote(Result.success(variations)))
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
+            flowOf(VariationsResult.Remote(Result.success(variations)))
         )
-        whenever(variationsDataSource.canLoadMore(any())).thenReturn(false)
-        whenever(variationsDataSource.loadMore(any())).thenReturn(Result.success(emptyList()))
+        whenever(variationsDataSource.canLoadMoreVariations(any())).thenReturn(false)
+        whenever(variationsDataSource.loadMoreVariations(any())).thenReturn(Result.success(emptyList()))
 
         val viewModel = createViewModel()
         advanceUntilIdle()
@@ -293,12 +293,12 @@ class WooPosVariationsViewModelTest {
         val variations = listOf(
             ProductTestUtils.generateProductVariation(1, 1, "10.0").toWooPosVariation(mapper),
         )
-        whenever(variationsDataSource.loadMore(any())).thenReturn(Result.success(variations))
-        whenever(variationsDataSource.canLoadMore(any())).thenReturn(true)
-        whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
+        whenever(variationsDataSource.loadMoreVariations(any())).thenReturn(Result.success(variations))
+        whenever(variationsDataSource.canLoadMoreVariations(any())).thenReturn(true)
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
             flow {
                 emit(
-                    FetchResult.Remote(
+                    VariationsResult.Remote(
                         Result.success(
                             listOf(
                                 ProductTestUtils.generateProductVariation(1, 1, "10.0").toWooPosVariation(mapper),
@@ -326,12 +326,12 @@ class WooPosVariationsViewModelTest {
     @Test
     fun `given load more fails, when end of list reached, then pagination state is error`() = runTest {
         // GIVEN
-        whenever(variationsDataSource.loadMore(any())).thenReturn(Result.failure(Exception()))
-        whenever(variationsDataSource.canLoadMore(any())).thenReturn(true)
-        whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
+        whenever(variationsDataSource.loadMoreVariations(any())).thenReturn(Result.failure(Exception()))
+        whenever(variationsDataSource.canLoadMoreVariations(any())).thenReturn(true)
+        whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
             flow {
                 emit(
-                    FetchResult.Remote(
+                    VariationsResult.Remote(
                         Result.success(
                             listOf(
                                 ProductTestUtils.generateProductVariation(1, 1, "10.0").toWooPosVariation(mapper),
@@ -363,8 +363,8 @@ class WooPosVariationsViewModelTest {
                 ProductTestUtils.generateProductVariation(1, 1, "10.0").toWooPosVariation(mapper),
                 ProductTestUtils.generateProductVariation(2, 1, "20.0").toWooPosVariation(mapper)
             )
-            whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
-                flowOf(FetchResult.Remote(Result.success(variations)))
+            whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
+                flowOf(VariationsResult.Remote(Result.success(variations)))
             )
 
             // WHEN
@@ -384,8 +384,8 @@ class WooPosVariationsViewModelTest {
     fun `given variation clicked and source is product list, when item clicked, then sends event with product list source`() =
         runTest {
             // GIVEN
-            whenever(variationsDataSource.fetchFirstPage(any(), any())).thenReturn(
-                flowOf(FetchResult.Remote(Result.success(emptyList())))
+            whenever(variationsDataSource.fetchVariationsFirstPage(any(), any())).thenReturn(
+                flowOf(VariationsResult.Remote(Result.success(emptyList())))
             )
 
             val viewModel = createViewModel()
@@ -413,8 +413,8 @@ class WooPosVariationsViewModelTest {
     fun `given search results screen, when variation clicked, then sends event with search results as source`() =
         runTest {
             // GIVEN
-            whenever(variationsDataSource.fetchFirstPage(anyOrNull(), anyOrNull())).thenReturn(
-                flowOf(FetchResult.Remote(Result.success(emptyList())))
+            whenever(variationsDataSource.fetchVariationsFirstPage(anyOrNull(), anyOrNull())).thenReturn(
+                flowOf(VariationsResult.Remote(Result.success(emptyList())))
             )
             val viewModel = createViewModel(123L, sourceType = ItemsListSourceType.SEARCH_RESULT)
             advanceUntilIdle()
